requirements.txt
asn1crypto==0.24.0
autopep8==1.4.3
bcrypt==3.1.5
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonpickle==1.0
MarkupSafe==1.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
